pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude=["*tests*", "*examples*", "*htmlcov*"]

[tool.setuptools.package-data]
static = ["hipposerve/web/static/*.svg", "hipposerve/web/templates/*.html"]

[project]
name = "hipposerve"
version = "0.0.1a7"
requires-python = ">=3.11"
dependencies = [
    "pydantic",
    "pydantic-settings",
    # 4.1.0 and above introduce a deprecation warning
    # that breaks the latest versions of asyncer.
    "anyio<4.1.0",
    "asyncer",
    "fastapi",
    "uvicorn",
    "beanie",
    "minio",
    "jinja2",
    "xxhash",
    "jinja-markdown",
    "typer",
    "httpx",
    "astropy",
    "pyjwt",
    "pwdlib[argon2]",
    "python-multipart",
    "loguru",
    "tqdm",
    "textual[syntax]"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "testcontainers",
    "coverage",
    "pytest-cov",
    "pytest-asyncio",
    "pytest-xprocess"
]

[project.scripts]
henry = "hippoclient.cli:main"

[tool.ruff.lint]
extend-select = ["I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
source = [
    "hipposerve",
    "hippoclient"
]
omit = [
    "hipposerve/web/*"
]
